Auto await feature (#186)
* WIP: (Refactoring) Implemented the future
boilerplate in the rendering stage. Just render
and test (missing the parse stage).

* WIP: removed `ReplaceFutureAttribute` hack in
`fixture` and moved the parsing code in
`extend_with_function_attrs` trait impl

* Move test in the right module and simplfied it

* Add check also for no future args

* Refactored and removed useless tests
that're already tested in future module

* Use Arguments also in tests and removed
the old impl. Also recoverd tests for
errors but implemented them in future
module

* WIP: Implementing await policy

* WIP: implemented the render part.

* WIP: implemented tests for is_future_await
logic

* WIP: Implemented future options parsing.
Missed the global one

* WIP implemented all global parser,
pass arguments info down to the
renderers and impleneted all tests
in render code to check that the
parsed info will used correctly.

* Changed impl: we should await in the original
method because we want the future in the
signature. Now E2E tests for fixture are
enabled and work.

* Integration tests

* Add documentation and changelog info