feat(test_attr): write `fn resolve_test_attr` and apply through the code
This function emits the proper test attribute as required, or raises
an error if the test is async and no explicit or implicit test attribute
was found.

Because it may emit an error, several additional functions needed
to also potentially emit errors, so that they could be collected
at the top level.