feat(test_attr): write `fn resolve_test_attr`
This function emits the proper test attribute as required, or raises
an error if the test is async and no explicit or implicit test attribute
was found.

This emits a compile error where the test is async and not explicit
or implicit test attribute was found. However, this mode of error gives
a relatively poor developer experience, so we'll want to add an
appropriate error check for this case in the error-handling section
of the macro as well.