test(test_attr): add tests of explicit and implicit test attr use
Also fix docs and usage where necessary.

Note that during the process of testing it was determined that we
cannot support the trailing-exclamation format recommended
by `macro_rules_attribute::apply`; that can't be parsed by `syn`,
and it's not worth using a different parser for this case.
That's ok, as `macro_rules_attribute::apply` appears to be willing
to (begrudgingly) accept input lacking the trailing exclamation.