docs(test_attr): test attributes should bind the test not the case
The requirement that test attributes bind to the test and not the
case was previously undocumented. This is much more relevant now
that test attributes must be specified in the async case.

However, this is _not_ a change in behavior; it always acted like
this. It just wasn't written, previously.